# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trust Giving - Charity Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Cause {
  id: number;
  title: string;
  description: string;
  month: number;
  year: number;
  goalAmount: number;
  raisedAmount: number;
  /** @nullable */
  imageUrl: string | null;
  category: string;
  isCurrent: boolean;
  /** @nullable */
  impact?: string | null;
  /** @nullable */
  beneficiaries?: string | null;
}

export interface Donor {
  id: number;
  name: string;
  amount: number;
  month: number;
  year: number;
  causeTitle: string;
  isAnonymous: boolean;
  /** @nullable */
  message?: string | null;
  donatedAt?: string;
}

export interface DonationInput {
  donorName: string;
  email: string;
  amount: number;
  causeId: number;
  /** @nullable */
  message?: string | null;
  isAnonymous?: boolean;
  /** @nullable */
  phone?: string | null;
}

export interface DonationReceipt {
  receiptId: number;
  receiptNumber: string;
  donorName: string;
  email: string;
  amount: number;
  causeTitle: string;
  causeDescription?: string;
  donatedAt: string;
  /** @nullable */
  message?: string | null;
  trustName?: string;
  taxExemptStatus?: string;
}

export interface StatsSummary {
  totalDonations: number;
  totalDonors: number;
  totalRaised: number;
  currentMonthRaised: number;
  currentCauseTitle: string;
  activeCauses?: number;
}

export interface MonthlyStatItem {
  month: number;
  year: number;
  totalRaised: number;
  donorCount: number;
  causeTitle: string;
}

export type ListDonorsParams = {
  month?: number;
  year?: number;
};