# lib/api-zod/src/generated/types/statsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trust Giving - Charity Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface StatsSummary {
  totalDonations: number;
  totalDonors: number;
  totalRaised: number;
  currentMonthRaised: number;
  currentCauseTitle: string;
  activeCauses?: number;
}